fix: assure proper typings _inputNode across lion fields
