+++ packages/fieldset/test/lion-fieldset.test.js
@@ -1,4 +1,6 @@
 import { runFormGroupMixinSuite } from '@lion/form-core/test-suites/form-group/FormGroupMixin.suite.js';
+import { runFormGroupMixinInputSuite } from '@lion/form-core/test-suites/form-group/FormGroupMixin-input.suite.js';
 import '../lion-fieldset.js';
 
 runFormGroupMixinSuite({ tagString: 'lion-fieldset' });
+runFormGroupMixinInputSuite({ tagString: 'lion-fieldset' });
